cdk/requirements.txt
boto3
aws-cdk.core
aws_cdk.aws_iam
aws_cdk.aws_s3
aws_cdk.aws_s3_deployment
aws_cdk.aws_cloudfront
aws_cdk.aws_route53
aws_cdk.aws_route53_targets